(maint) Refactor sensu_check_config with spec tests
Without this patch there are no rspec examples of how providers should behave.
This is a problem because there should be a reference to build from when
changing the providers.

This patch addresses the problem by adding rspec coverage for the
sensu_check_config json provider.

Note: Garrett and I decided to remove the sensu_check_config type and provider.
It doesn't operate as expected in the master branch, is not used anywhere in the
module, and the custom configuration map in sensu_check provides the same
behavior for checks.  Each specific type will implement the generic
configuration map similar to sensu_check.

The reference implementation of the spec tests will move to sensu_check in a
subsequent patch.  sensu_check_config will be removed in a subsequent patch.

Relates to #764